+++ src/Notify.ts
@@ -2,6 +2,9 @@ import { NotificationManager } from "./notification-manager";
 
 const globalManager = new NotificationManager();
 
+/**
+ * @deprecated Use `snackbar()` instead
+ */
 const notify: (notification: {
     message: string;
     duration?: number;
@@ -16,6 +19,20 @@ const notify: (notification: {
     classes?: Array<string> | string;
 }) => void = globalManager.notify.bind(globalManager);
 
+const snackbar: (notification: {
+    message: string;
+    duration?: number;
+    closeable?: boolean;
+    buttons?: Array<{
+        label: string;
+        callback: Function;
+        ariaLabel?: string;
+        classes?: Array<string> | string;
+    }>;
+    force?: boolean;
+    classes?: Array<string> | string;
+}) => void = globalManager.snackbar.bind(globalManager);
+
 const toast: (notification: {
     title: string;
     message: string;
@@ -26,4 +43,4 @@ const toast: (notification: {
     element?: HTMLElement;
 }) => void = globalManager.toast.bind(globalManager);
 
-export { NotificationManager, notify, toast };
+export { NotificationManager, notify, toast, snackbar };